moved files to Model
