+++ Application/src/model/DiagramCheck.java
@@ -1,3 +1,5 @@
+package model;
+
 import javafx.scene.control.Alert;
 import java.util.Collection;
