+++ Application/src/model/Import.java
@@ -1,3 +1,5 @@
+package model;
+
 import javafx.scene.control.Alert;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;
